==> methyl/01_pval_plot.ipynb <==
library(ggplot2)
library(reshape2)
library(dplyr)
library(data.table)
# library(tidyverse)

col <- c('#FF204E', '#00224D', '#2B3499')
# col <- c('#FFD369', '#393E46', '#222831')
# col <- c('#F266AB', '#2CD3E1', '#E74646')

path <- '01_organ/18_250320'

# path <- 'S448'
# minfi <- read.table('/disk0/sm/methyl/S448/table_QC.txt', sep='\t', header=T)
# head(minfi)
# minfi <- minfi[order(minfi$File_Name),]
# # minfi <- minfi[order(minfi$X),] %>% filter(startsWith(X, 'L'))
# minfi$n <- seq(1, nrow(minfi))
# m_minfi <- melt(minfi, id.vars='n', measure.vars=c('True_rate', 'False_rate'))
# m_minfi$variable <- factor(m_minfi$variable, levels=c('False_rate', 'True_rate'))
# head(m_minfi, 2)

minfi <- data.frame(read.table(paste0('/disk0/sm/methyl/', path, '/data/pval_minfi.txt'), sep='\t', header=T, row.names = NULL))
names(minfi)[1] <- 'X'
m_minfi <- reshape2::melt(minfi, id.vars='X', measure.vars=c('True_rate', 'False_rate'))
m_minfi$variable <- factor(m_minfi$variable, levels=c('False_rate', 'True_rate'), labels=c("0.05 초과", '0.05 이하'))

p <- ggplot(data=m_minfi, aes(x=X, y=value, fill=variable)) +
geom_bar(stat='identity', position="stack") +
# coord_fixed(ratio = 0.5) +
  coord_cartesian(ylim = c(98, 100))  +
  # scale_x_continuous(breaks = seq(0, ) +
  scale_y_continuous(breaks = seq(98, 100, 0.2)) +
  labs(title = "Detection P-value", 
        x = 'plate', 
        y = 'detection p-value < 0.05 marker rate') + 
  theme_bw() +
  theme(
  #   axis.text.x = element_text(seq(0, 2198), angle=30, hjust=1),
    axis.text.x = element_blank(),
    axis.text.y = element_text(size=10),
    plot.title = element_text(hjust = 0.5,size=12,face='bold'),
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.ticks.x = element_blank(),
    legend.text = element_text(size=10),
    legend.position = 'right',
    panel.grid.major.x = element_blank(), 
    panel.grid.major.y = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_blank(), 
    legend.title = element_blank()
    ) +
  geom_hline(yintercept=99.5, color='red') +
  scale_fill_manual(values=c(col[1], col[2])) #c('#FF204E', '#00224D')
# ggsave(p, file='/disk2/kb/sm/chip/plot/final/total.png', width=2833, height=1890, units='px')
ggsave(p, file=paste0('/disk0/sm/methyl/', path, '/plot/pval_minfi.png'), width=3500, height=3000, units='px', dpi=500)

p <- ggplot(minfi, aes(x=True_rate)) + 
  coord_cartesian(xlim = c(99, 100)) +
  geom_histogram(fill=col[3], binwidth = 0.01) + 
  scale_x_continuous(breaks = seq(99, 100, 0.1)) +
  labs(title = "Detection P-value", 
      x = 'p-value', 
      y = 'Number of samples') +
  # scale_y_continuous(breaks = seq(90, 100, 2)) +
  theme_bw() +
  theme(
  #   axis.text.x = element_text(seq(0, 2198), angle=30, hjust=1),
    axis.text.x = element_text(size=8),
    axis.text.y = element_text(size=10),
    plot.title = element_text(hjust = 0.5,size=12,face='bold'),
    axis.title.x = element_text(size=10),
    axis.title.y = element_text(size=10),
    axis.ticks.x = element_blank(),
    legend.text = element_text(size=10),
    legend.position = 'none',
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey80"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "gray80"), 
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_blank(), 
    legend.title = element_blank()
    ) +
  geom_vline(xintercept=99.5, color='red')
ggsave(p, file=paste0('/disk0/sm/methyl/', path, '/plot/pval_histo.png'), width=3000, height=3000, units='px', dpi=500)

# gene <- read.table('/disk0/sm/methyl/03_pollution/total/data/gene_count.txt', sep='\t', header=T)
# gene <- gene[gene$count > 2, ]
# head(gene, 2)

# p <- ggplot(gene, aes(x=reorder(Gencode, -count), y=count))+
#   geom_bar(stat = "identity", size=.2, width=0.6, fill='#FB8072') +
#   labs(x='gene', y = 'number of cpg') +
#   theme_bw() +
#   theme(
#     axis.text.x = element_text(size=7, angle=90, hjust=1),
#     axis.text.y = element_text(size=5),
#     plot.title = element_text(hjust = 0.5,size=10,face='bold'),
#     axis.title.x = element_text(size=8),
#     axis.title.y = element_text(size=8),
#     legend.text = element_text(size=8),
    
#     panel.grid.major.x = element_line(linetype = "dotted", colour = "grey70"),
#     panel.grid.major.y = element_line(linetype = "dotted", colour = "grey70"),
#     panel.grid.minor.x = element_blank(), 
#     panel.grid.minor.y = element_line(linetype = "dotted", colour = "grey70"),
#     legend.title = element_blank(),
#     legend.position = 'none'
#         ) +
#     scale_fill_manual(values=c('#FB8072'))
# ggsave(p, file='/disk0/sm/methyl/03_pollution/total/plot/gene.png', width=8000, height=6000, units='px', dpi=1000)
