==> methyl/02_densityplot.ipynb <==
library(minfi)
library(dplyr)
library(RColorBrewer)
library(ggplot2)
library(viridis)
library(reshape2)
library(data.table)

library(doParallel)

registerDoParallel(cores=24)

col <- c('#FF6C22', '#000B58')
# col <- c('#FF204E', '#046B99')
# col <- c('#FFD369', '#739072')
col <- c('#FF6C22', '#046B99')

data <- fread("/home/data/two-way_normalized_beta_value_all_CpGs.txt", sep="\t", fill=TRUE, header=TRUE)
# data <- fread("/home/data/Noob_beta_value_all_CpGs.txt", sep="\t", fill=TRUE, header=TRUE)
# data[1:4, 1:4]
dim(data)

png(filename="/home/data/plot/two-way_densityplot.png", width=800, height=500)
densityPlot(as.matrix(data[, 2:-1]), sampGroups = NULL, main = "", xlab = "Beta", pal = col[1], legend = TRUE)
dev.off()

# raw <- read.table("/home/data/raw_beta_value_all_CpGs.txt", sep="\t", fill = TRUE, header=TRUE, row.names=1)
raw <- fread("/home/data/raw_beta_value_all_CpGs.txt", sep="\t", fill = TRUE, header=TRUE)
head(raw)
dim(raw)


png(filename="/home/data/plot/raw_densityplot.png", width=800, height=500)
densityPlot(as.matrix(raw[, 2:-1]), sampGroups = NULL, main = "", xlab = "Beta", pal = col[2], legend = FALSE)
dev.off()

S2023 <- fread("/home/data/2023/01_1760/two-way_normalized_beta_value_all_CpGs.txt", sep="\t", fill=TRUE, header=TRUE)
S2024 <- fread("/home/data/2024/03_pollution/total/two-way_normalized_beta_value_all_CpGs.txt", sep="\t", fill=TRUE, header=TRUE)
data <- fread("/home/data/S448/two-way_normalized_beta_value_all_CpGs.txt", sep="\t", fill=TRUE, header=TRUE)

a <- S2023[,2:length(S2023)]
b <- S2024[,2:length(S2024)]
beta <- cbind(a, b)
group <- c(rep('2023', length(a)), rep('2024', length(b)))
group <- factor(group, levels=c('2023', '2024'))
png(filename=paste0("/home/data/", path, "/plot/ab_densityplot.png"), width=800, height=500, unit="px")
densityPlot(as.matrix(beta), sampGroups = group, main = "", xlab = "Beta", pal = c(col[2], adjustcolor(col[1],alpha.f = 0.1)), add = TRUE, legend = TRUE)
dev.off()
