==> methyl/03_age_plot.ipynb <==
# if (!require("BiocManager", quietly = TRUE))
#           install.packages("BiocManager")

# BiocManager::install("methylclock", force = TRUE)
# BiocManager::install("impute")

# install.package(tibble)
# install.package(ggplot2)
# install.package(ggpmisc)
# install.package(GEOquery)

# library(methylclockData)
# library(methylclock)
# library(Biobase)
# library(tibble)
# library(impute)
# library(ggplot2)
# library(ggpmisc)
# library(GEOquery)
# library(data.table)

# MethylationData <- fread('./two-way_normalized_beta_value_all_CpGs.txt', header=TRUE, sep = '\t')
# MethylationData <- data.frame(norm_by_bmiq_with_name)
# # MethylationData <- read.csv('/disk0/sm/methyl/03_pollution/total/two-way_normalized_beta_value_all_CpGs.txt', header=TRUE, sep = '\t')
# MethylationData$cpgs <- sub("_.+$", "", MethylationData$cpgs)
# head(MethylationData$cpgs)

# checkClocksGA(MethylationData)

# age.example55 <- DNAmAge(MethylationData)
# age.example55
# write.table(age.example55, file='./DNAmAge_Result.txt', sep='\t', row.names = F)

# age.example55.2 <- DNAmGA(MethylationData)
# age.example55.2
# write.table(age.example55.2, file='./DNAmGA_Result.txt', sep='\t', row.names = F)

library(ggplot2)
library(reshape)
library(RColorBrewer)
library(stringr)
library(dplyr)
library(reshape2)
library(grid)

library(doParallel)

registerDoParallel(cores=24)

age_normal <- read.table("/disk0/sm/methyl/03_pollution/total/data/age.txt", sep="\t", fill = TRUE, header=TRUE)
head(age_normal)

axis_num <- 'black'

# x=reorder(Name, Horvath)
p <- ggplot(data=age_normal, aes(x=reorder(Name, Horvath), y=Horvath, group=phase, color=phase)) + #reorder(id, value)
geom_smooth(size=0.5, level=0.999) +#, color=adjustcolor('red', alpha.f =0.5)) +
geom_point(size=0.5) +
coord_fixed(ratio = 0.5) +
  coord_cartesian(ylim = c(40, 110))  +
#   scale_x_continuous(breaks = seq(0, nrow(all_callrate), nrow(all_callrate)/plate_num)) +
  scale_y_continuous(breaks = seq(40, 110, 10)) +
  labs(title = "Horvath", 
        x = 'sample', 
        y = 'age difference') + 
  theme_bw() +
  theme(
  #   axis.text.x = element_text(seq(0, 2198), angle=30, hjust=1),
    axis.text.x = element_blank(),
    axis.text.y = element_text(size=18),
    plot.title = element_text(hjust = 0.5,size=20,face='bold'),
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.ticks.x = element_blank(),
    legend.text = element_text(size=18),
    legend.position = 'none',
    
    panel.grid.major.x = element_blank(), 
    panel.grid.major.y = element_line(linetype = "dotted", colour = "grey80"),
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_blank(), 
    legend.title = element_blank()
    ) +
  # geom_hline(yintercept=c(mean(age_1_11$Horvath), mean(age_1_11$Hannum), mean(age_1_11$Levine), mean(age_1_11$skinHorvath), mean(age_1_11$BLUP), mean(age_1_11$EN)), color=c('#FB8072', '#FDB462', '#FFED6F', '#B3DE69', '#80B1D3', '#BC80BD')) + 
  scale_colour_manual(values=c('#3A4D39', '#FF9209')) 
ggsave(p, file='/disk0/sm/methyl/03_pollution/total/plot/age_Horvath_check.png', width=6000, height=6000, units='px', dpi=1000)

age <- read.table("/disk0/sm/methyl/03_pollution/total/data/age_1_11.txt", sep="\t", fill = TRUE, header=TRUE)

age <- age[order(age$Horvath), ]
# head(age)
age$id <- factor(age$Name, level=age$Name)

m_age = melt(age)
# head(m_age)
summary(age)

colr =c('#FB8072', '#FDB462', '#FFED6F', '#B3DE69', '#80B1D3', '#BC80BD', '#ff6dbd', '#adadad')

m_age <- m_age %>% filter((variable == 'Horvath') | (variable == 'Levine') | (variable == 'BLUP') | (variable == 'SM'))

p <- ggplot(m_age, aes(x=variable, y=value, fill=variable)) + 
  # coord_cartesian(ylim = c(5, 35)) +
  geom_hline(yintercept=20, color='#E74646', size=0.7, linetype=2) +
  geom_boxplot(outlier.shape = NA) + #
  labs(
        x = 'age predict method', 
        y = 'age difference') + 
  # scale_fill_manual(values=c('#FB8072', '#FFED6F', '#ff6dbd', 'white')) +
  scale_fill_manual(values=colr) +
  theme_bw() +
  theme(
  #   axis.text.x = element_text(seq(0, 2198), angle=30, hjust=1),
    axis.text.x = element_text(size=15),
    axis.text.y = element_text(size=15),
    plot.title = element_text(hjust = 0.5,size=20,face='bold'),
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.ticks.x = element_blank(),
    legend.text = element_text(size=18),
    legend.position = 'none',
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey80"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "gray80"), 
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_blank(), 
    legend.title = element_blank()
    ) 
  # geom_jitter(shape=1, position=position_jitter(0.3), size=0.1) +
  #  geom_boxplot(width=0.1, outlier.shape = NA) 


  ggsave(p, file='/disk0/sm/methyl/03_pollution/total/plot/age_1_11_box.png', width=10000, height=10000, units='px', dpi=1000)
# dev.off()

# colr <- c('#FB8072', '#FDB462', '#FFED6F', '#B3DE69', '#80B1D3', '#BC80BD')

# p <- ggplot(m_age, aes(x=variable, y=value, fill=variable)) + 
#   coord_cartesian(ylim = c(5, 35)) +
#   geom_violin() + #
#   geom_boxplot(width=0.1, outlier.shape = NA) +
#   labs(title = "1 11", 
#         # x = 'algorithm', 
#         y = 'age difference') + 
#   scale_fill_manual(values=colr) +
#   theme_bw() +
#   theme(
#   #   axis.text.x = element_text(seq(0, 2198), angle=30, hjust=1),
#     axis.text.x = element_text(size=18),
#     axis.text.y = element_text(size=18),
#     plot.title = element_text(hjust = 0.5,size=20,face='bold'),
#     axis.title.x = element_blank(),
#     axis.title.y = element_blank(),
#     axis.ticks.x = element_blank(),
#     legend.text = element_text(size=18),
#     legend.position = 'none',
    
#     panel.grid.major.x = element_line(linetype = "dotted", colour = "grey80"),
#     panel.grid.major.y = element_line(linetype = "dotted", colour = "gray80"), 
#     panel.grid.minor.x = element_blank(), 
#     panel.grid.minor.y = element_blank(), 
#     legend.title = element_blank()
#     )
#   # geom_jitter(shape=1, position=position_jitter(0.3), size=0.1) +
#   #  geom_boxplot(width=0.1, outlier.shape = NA) 


#   ggsave(p, file='/disk0/sm/methyl/03_pollution/total/plot/age_1_11_violin.png', width=10000, height=10000, units='px', dpi=1000)
# # dev.off()

age <- read.table("/disk0/sm/methyl/03_pollution/total/data/age.txt", sep="\t", fill = TRUE, header=TRUE)
# head(age_normal)
# age <- age_normal[age_normal$phase == '1기',]
dim(age)

p <- ggplot(age, aes(x=Horvath, group=phase, fill=phase)) + 
#   coord_cartesian(xlim = c(99, 100)) +
  geom_histogram(binwidth = 1, position='identity') + 
#   scale_x_continuous(breaks = seq(99, 100, 0.1)) +
  # scale_y_continuous(breaks = seq(90, 100, 2)) +
  labs(x='Age distribution', y='Estimated age by Horvath algorithm')
  theme_bw() +
  theme(
  #   axis.text.x = element_text(seq(0, 2198), angle=30, hjust=1),
    axis.text.x = element_text(size=12),
    axis.text.y = element_text(size=15),
    plot.title = element_text(hjust = 0.5,size=20,face='bold'),
    axis.title.x = element_text(size=15),
    axis.title.y = element_text(size=15),
    axis.ticks.x = element_blank(),
    legend.text = element_text(size=18),
    legend.position = 'none',
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey80"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "gray80"), 
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_blank(), 
    legend.title = element_blank()
    ) +
    scale_fill_manual(values=adjustcolor(c('#3A4D39', '#FF9209'),alpha.f = 0.5)) 

ggsave(p, file='/disk0/sm/methyl/03_pollution/total/plot/age_histo.png', width=6000, height=6000, units='px', dpi=1000)

###### 그냥 양상 한번 봄 ## 아래는 실제 나이와 예측 나이를 비교한 것임

age <- read.table('/disk0/sm/methyl/03_pollution/total/data/age_table.txt', sep='\t', header=T)
# age$rate = age$EN/age$Actual.age
head(age, 2)
# dim(age)

algorithm <- names(age)[5:13]
for (i in algorithm){
  tmp <- age[c('Sample_Name', 'Age', i)]
  names(tmp) <- c('Sample_Name', 'Age', 'predict')
  print(paste(i, " : ", cor(tmp$Age, tmp$predict)))
  # print(paste(i, " : ", var(tmp$Age-tmp$predict)))
}

algorithm <- names(age)[5:13]
for (i in algorithm){
  tmp <- age[c('Sample_Name', 'Age', i)]
  names(tmp) <- c('Sample_Name', 'Age', 'predict')
  grob3 <- grobTree(textGrob(paste("Coefficient : ", cor(tmp$Age, tmp$predict)), x = 0.03, y = 0.97, hjust = 0, gp = gpar(col = "#ff4848", fontsize = 17, fontface = "bold")))
  p <- ggplot(data=tmp, aes(x=Age, y=predict)) +
  coord_cartesian(xlim = c(40, 90), ylim = c(20, 100))  +
  scale_x_continuous(breaks = seq(40, 90, 10)) +
  scale_y_continuous(breaks = seq(20, 100, 10)) +
  geom_point(size=0.5) +
    labs(title = i, 
          x = 'actual age', 
          y = 'predicted age') + 
    theme_bw() +
    theme(
    #   axis.text.x = element_text(seq(0, 2198), angle=30, hjust=1),
      axis.text.x = element_text(size=15),
      axis.text.y = element_text(size=15),
      plot.title = element_text(hjust = 0.5,size=20,face='bold'),
      axis.title.x = element_text(size=18),
      axis.title.y = element_text(size=18),
      axis.ticks.x = element_blank(), 
      legend.text = element_text(size=10),
      
      panel.grid.major.x = element_line(linetype = "dotted", colour = "grey80"),
      panel.grid.major.y = element_line(linetype = "dotted", colour = "gray70"), 
      panel.grid.minor.x = element_blank(), 
      panel.grid.minor.y = element_blank(), 
      legend.title = element_blank()
      ) +
      geom_smooth(method=lm, se=F, color='#ff0055') +
      annotation_custom(grob3)
  #   scale_colour_manual(values=c('#ff4040')) 
  # ggsave(p, file='/disk0/sm/chip/plot/final/total.png', width=2833, height=1890, units='px')
  ggsave(p, file=paste0('/disk0/sm/methyl/03_pollution/total/plot/age_', i, '.png'), width=8000, height=8000, units='px', dpi=1000)
}

age_final <- read.table("/disk0/sm/methyl/03_pollution/total/data/age_algorithm.txt", sep="\t", fill = TRUE, header=TRUE)

# age_final <- age_final[order(age_final$Horvath), ]
# head(age_final)
# age_final$id <- factor(age_final$id, level=age_final$id)

m_age_final = melt(age_final)
head(m_age_final, 2)
# summary(age_final)
# m_age_final <- m_age_final %>% filter(variable != 'SM')

# m_age_final <- m_age_final %>% filter((variable == 'Horvath') | (variable == 'Levine') | (variable == 'BLUP') | (variable == 'SM'))

p <- ggplot(m_age_final, aes(x=variable, y=value, fill=variable)) + 
  # coord_cartesian(ylim = c(-50, 50)) +
  geom_hline(yintercept=0, color='red', linetype = 2) +
  geom_boxplot(outlier.shape = NA) + #
  labs(
        x = 'age predict method', 
        y = 'age difference') + 
  # scale_fill_manual(values=c('#FB8072', '#FFED6F', '#ff6dbd', 'white')) +
  scale_fill_manual(values=c(colr, 'white')) +
  theme_bw() +
  theme(
    axis.text.x = element_text(size=15),
    axis.text.y = element_text(size=15),
    # plot.title = element_text(hjust = 0.5,size=10,face='bold'),
    axis.title.x = element_text(size=18),
    axis.title.y = element_text(size=18),
    axis.ticks.x = element_blank(), 
    legend.text = element_blank(), 
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_line(linetype = "dotted", colour = "grey70"),
    legend.title = element_blank(),
    legend.position = 'none',
  ) 
 
  # geom_jitter(shape=1, position=position_jitter(0.3), size=0.1) +
  #  geom_boxplot(width=0.1, outlier.shape = NA) 
  


  ggsave(p, file='/disk0/sm/methyl/03_pollution/total/plot/age_final.png', width=10000, height=10000, units='px', dpi=1000)
# dev.off()
