==> methyl/06_gene_plot.ipynb <==
# library(minfi)
library(dplyr)
library(RColorBrewer)
library(ggplot2)
library(viridis)
library(reshape2)
library(data.table)

library(doParallel)

registerDoParallel(cores=24)

# library(pheatmap) ## for heatmap generation
# # library(tidyverse) ## for data wrangling
# library(ggplotify) ## to convert pheatmap to ggplot2
# library(heatmaply) ## for constructing interactive heatmap

f <- data.frame(t(beta['cg10270053_TC21', 1:1720]))
f$group <- c(rep('1', 860), rep('11', 860))
names(f) <- c('ID', 'group')
# beta <- cbind(pheno, beta)
# beta <- data.frame(beta)
colr <- c('#ffd45d', '#1f86fc')



p <- ggplot(f, aes(x=group, y=ID, group=group, fill=group)) + 
  geom_boxplot(outlier.shape = NA) + #
  scale_fill_manual(values=colr) +
  labs(title='cg10270053_TC21', y='beta') + 
  theme_bw() +
  theme(
    axis.text.x = element_text(size=5),
    axis.text.y = element_text(size=5),
    plot.title = element_text(hjust = 0.5,size=10,face='bold'),
    axis.title.x = element_blank(),
    axis.title.y = element_text(size=8),
    legend.text = element_text(size=8),
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_line(linetype = "dotted", colour = "grey70"),
    legend.title = element_blank(),
    legend.position = 'right'
  )
ggsave(p, file='/disk2/kb/sm/methyl/meth/01/plot/cpg_esr1.png', width=4000, height=6000, units='px')


# gene <- read.table('/disk0/sm/methyl/03_pollution/total/data/gene_count.txt', sep='\t', header=T)
gene <- read.table('/disk0/sm/methyl/03_pollution/total/data/dmp_gene.txt', sep='\t', header=T)
gene <- gene[gene$count > 11, ]
# gene_s <- 
head(gene)
dim(gene)

p <- ggplot(gene, aes(x=reorder(G11_to_G1.gene, -count), y=count))+
  geom_bar(stat = "identity", width=0.6, fill='#FB8072') +
  labs(x='gene', y = 'number of cpg') +
  theme_bw() +
  theme(
    axis.text.x = element_text(size=10, angle=90, hjust=1),
    axis.text.y = element_text(size=15),
    plot.title = element_text(hjust = 0.5,size=10,face='bold'),
    axis.title.x = element_blank(), 
    axis.title.y = element_text(size=18),
    legend.text =element_blank(), 
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_line(linetype = "dotted", colour = "grey70"),
    legend.title = element_blank(),
    legend.position = 'none'
        ) +
    scale_fill_manual(values=c('#FB8072'))
ggsave(p, file='/disk0/sm/methyl/03_pollution/total/plot/gene.png', width=8000, height=4000, units='px', dpi=1000)

island <- read.table('/disk2/kb/sm/methyl/meth/01/island_count.txt', sep='\t', header=T)
island$Name <- c('CpG OpenSea', 'CpG Shore', 'CpG Island', 'CpG Shelf')
island$Name <- factor(island$Name , level=c('CpG Island', 'CpG Shore', 'CpG Shelf', 'CpG OpenSea'))

head(island)
dim(island)

p <- ggplot(island, aes(x=Name, y=count, fill=Name))+
  geom_bar(stat = "identity", size=.2, width=0.6) +
  coord_cartesian(ylim = c(0, 450000)) +
  scale_y_continuous(breaks = seq(0, 450000, 50000), labels = c('0', '50000', '100000', '150000', '200000', '250000', '300000', '350000', '400000', '450000')) +
  labs(x='CpG Annotations', y = 'number of DMR') +
  theme_bw() +
  theme(
    axis.text.x = element_text(size=7),
    axis.text.y = element_text(size=5),
    plot.title = element_text(hjust = 0.5,size=10,face='bold'),
    axis.title.x = element_text(size=8),
    axis.title.y = element_text(size=8),
    legend.text = element_text(size=8),
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey90"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "grey80"),
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_blank(),
    legend.title = element_blank(),
    legend.position = 'none'
        ) +
    scale_fill_manual(values=c('#fbc072', '#fbb472', '#79b3ff', '#3754f5'))
ggsave(p, file='/disk2/kb/sm/methyl/meth/01/plot/island.png', width=2000, height=1500, units='px')

age <- read.table('/disk2/kb/sm/methyl/meth/01/age_1720.txt', sep='\t', header=T)
age$rate = age$Horvath/age$Actual.age
head(age)
dim(age)

library(RColorBrewer)
library(dplyr)
library(reshape2)
library(ggplot2)
library(grid)
library(RColorBrewer)

axis_num <- 'black'
grob3 <- grobTree(textGrob(paste("Coefficient : ", cor(age$Actual.age, age$Horvath)), x = 0.03, y = 0.97, hjust = 0, gp = gpar(col = "#ff4848", fontsize = 7, fontface = "bold")))
p <- ggplot(data=age, aes(x=Actual.age, y=Horvath)) +
geom_point(size=0.1) +
  labs(title = "AGE Analysis", 
        x = 'actual age', 
        y = 'predicted age') + 
  theme_bw() +
  theme(
  #   axis.text.x = element_text(seq(0, 2198), angle=30, hjust=1),
    axis.text.x = element_text(size=5),
    axis.text.y = element_text(size=5),
    plot.title = element_text(hjust = 0.5,size=10,face='bold'),
    axis.title.x = element_text(size=8),
    axis.title.y = element_text(size=8),
    axis.ticks.x = element_line(color = axis_num),
    legend.text = element_text(size=10),
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey80"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "gray70"), 
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_blank(), 
    legend.title = element_blank()
    ) +
    geom_smooth(method=lm, se=F, color='#ff0055') +
    annotation_custom(grob3)
#   scale_colour_manual(values=c('#ff4040')) 
# ggsave(p, file='/disk2/kb/sm/chip/plot/final/total.png', width=2833, height=1890, units='px')
ggsave(p, file='/disk2/kb/sm/methyl/meth/01/plot/age_1720.png', width=1500, height=1500, units='px')

age <- read.table('/disk2/kb/sm/methyl/meth/01/age_strange.txt', sep='\t', header=T)
head(age)
dim(age)

axis_num <- 'black'
grob3 <- grobTree(textGrob(paste("Coefficient : ", cor(age$Actual.age, age$strange)), x = 0.03, y = 0.97, hjust = 0, gp = gpar(col = "#ff4848", fontsize = 7, fontface = "bold")))
p <- ggplot(data=age, aes(x=Actual.age, y=strange)) +
geom_point(size=0.1) +
  labs(title = "AGE Analysis", 
        x = 'actual age') + 
  theme_bw() +
  theme(
  #   axis.text.x = element_text(seq(0, 2198), angle=30, hjust=1),
    axis.text.x = element_text(size=5),
    axis.text.y = element_text(size=5),
    plot.title = element_text(hjust = 0.5,size=10,face='bold'),
    axis.title.x = element_text(size=8),
    axis.title.y = element_text(size=8),
    axis.ticks.x = element_line(color = axis_num),
    legend.text = element_text(size=10),
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey80"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "gray70"), 
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_blank(), 
    legend.title = element_blank()
    ) +
    geom_smooth(se=F, color='#ff0055') +
    annotation_custom(grob3)
#   scale_colour_manual(values=c('#ff4040')) 
# ggsave(p, file='/disk2/kb/sm/chip/plot/final/total.png', width=2833, height=1890, units='px')
ggsave(p, file='/disk2/kb/sm/methyl/meth/01/plot/age_strange.png', width=1500, height=1500, units='px')

# box <- read.table("/disk2/kb/sm/methyl/meth/01/data/dmpdmr_heat_111.txt", sep="\t", fill = TRUE, header=TRUE, row.names=1)[, 1:1720]
gender <- read.table("/disk2/kb/sm/methyl/meth/01/data/gender.txt", sep="\t", fill = TRUE, header=TRUE)[1:1720,]
# m_box <- melt(box)

install.packages("ggpubr")
library("ggpubr")
library("dplyr") 

f <- data.frame(t(beta['cg10270053_TC21', 1:1720]))
f$group <- c(rep('1', 860), rep('11', 860))
names(f) <- c('ID', 'group')
f$gender <- gender$gender

colr <- c('#5dffd7', '#fc1f8d')
#5dffbc

# p + 
p <- ggplot(f, aes(x=group, y=ID, fill=gender), stat = 'identity') + 
  geom_boxplot(outlier.shape = NA) + #
  scale_fill_manual(values=colr) +
  labs(title='ESR1', y='beta') + 
  theme_bw() +
  theme(
    axis.text.x = element_text(size=5),
    axis.text.y = element_text(size=5),
    plot.title = element_text(hjust = 0.5,size=10,face='bold'),
    axis.title.x = element_blank(),
    axis.title.y = element_text(size=8),
    legend.text = element_text(size=8),
    
    panel.grid.major.x = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.major.y = element_line(linetype = "dotted", colour = "grey70"),
    panel.grid.minor.x = element_blank(), 
    panel.grid.minor.y = element_line(linetype = "dotted", colour = "grey70"),
    legend.title = element_blank(),
    legend.position = 'right'
  ) +
  stat_compare_means(method = "t.test")
ggsave(p, file='/disk2/kb/sm/methyl/meth/01/plot/gender_esr1.png', width=1500, height=1500, units='px')
